refactor: AuthService 인터페이스 생성
